+++ lib/bluetooth_print_model.g.dart
@@ -39,8 +39,7 @@ LineText _$LineTextFromJson(Map<String, dynamic> json) {
     weight: json['weight'] as int?,
     width: json['width'] as int?,
     height: json['height'] as int?,
-    absolutePos: json['absolutePos'] as int?,
-    relativePos: json['relativePos'] as int?,
+    relativeX: json['relativeX'] as int?,
     fontZoom: json['fontZoom'] as int?,
     underline: json['underline'] as int?,
     linefeed: json['linefeed'] as int?,
@@ -65,8 +64,7 @@ Map<String, dynamic> _$LineTextToJson(LineText instance) {
   writeNotNull('weight', instance.weight);
   writeNotNull('width', instance.width);
   writeNotNull('height', instance.height);
-  writeNotNull('absolutePos', instance.absolutePos);
-  writeNotNull('relativePos', instance.relativePos);
+  writeNotNull('relativeX', instance.relativeX);
   writeNotNull('fontZoom', instance.fontZoom);
   writeNotNull('underline', instance.underline);
   writeNotNull('linefeed', instance.linefeed);